requirements.txt
django
djangorestframework
psycopg2-binary
djongo
pymongo
djangorestframework-simplejwt
django-environ
drf-yasg
mongoengine
django-injector
dotenv